Move project providers to root project folder
