+++ web/app/providers/CurrentProjectProvider.tsx
@@ -12,8 +12,8 @@ import { type StateType } from '~/lib/store'
 import { getItem, removeItem } from '~/utils/localstorage'
 import routes from '~/utils/routes'
 
-import { getProjectPreferences, setProjectPreferences } from '../View/utils/cache'
-import { CHART_METRICS_MAPPING } from '../View/ViewProject.helpers'
+import { getProjectPreferences, setProjectPreferences } from '../pages/Project/View/utils/cache'
+import { CHART_METRICS_MAPPING } from '../pages/Project/View/ViewProject.helpers'
 
 interface CurrentProjectContextType {
   id: string
